Roadmap 1.7: warning-free at /W3, /W4 and -Wall -Wextra on all four compilers
cmake/Warnings.cmake (was WarningsAsErrors.cmake) puts /W3 on MSVC, /W4 on clang-cl and
-Wall -Wextra on GCC and Clang on every first-party target, next to the existing
TASFW_WARNINGS_AS_ERRORS, which now passes at those levels on all four compilers (GCC 13
and Clang 17 checked in a Docker container before CI). clang-cl needs /W4 because a
GNU-style -Wall is taken as MSVC's /Wall, which is -Weverything (docs/compilers.md).

The inventory before the fixes: 106 MSVC sites at /W3, 126 clang-cl at -Wall, 155 GCC and
166 Clang at -Wall -Wextra. Nearly every fix is an explicit cast of the conversion that was
already happening, a deleted dead local or field, or an unnamed parameter. The rest: a
virtual destructor on Resource, snprintf for sprintf, int64_t for the tracked-state hooks'
ad-hoc levels and BitFsPyramidOscillation_Iteration's frames, RequireObject taking a
std::string_view (GCC 13 -Wdangling-reference on a temporary bound to a const
std::string&), the decomp's static helper declarations moved out of sm64/Surface.hpp
(-Wunused-function per translation unit), and BitFsScApproach_AttemptDr_BF losing an
unused constructor argument.

Exact counts identical on both compilers; the drift test still matches the DLL bit for bit.
Tier C and D flat on both once Docker's VM was stopped (it added 11.7% to the
deterministic Tier D run). An A/B against the pre-change tree attributes +0.6% to the
oscillation row and a clang-cl layout shift of +12 to +17% to the uncached GetInputs rows
(MSVC's moved -10%); those rows are re-saved, with the numbers in docs/performance.md.

+++ tasfw-core/src/core/Inputs.cpp
@@ -41,14 +41,14 @@ PopulateInputMappings()
 			float adjustedStickY = 0;
 
 			if (stickX <= -8)
-				adjustedStickX = stickX + 6;
+				adjustedStickX = float(stickX + 6);
 			else if (stickX >= 8)
-				adjustedStickX = stickX - 6;
+				adjustedStickX = float(stickX - 6);
 
 			if (stickY <= -8)
-				adjustedStickY = stickY + 6;
+				adjustedStickY = float(stickY + 6);
 			else if (stickY >= 8)
-				adjustedStickY = stickY - 6;
+				adjustedStickY = float(stickY - 6);
 
 			float stickMag = sqrtf(
 				adjustedStickX * adjustedStickX + adjustedStickY * adjustedStickY);
@@ -77,7 +77,7 @@ PopulateInputMappings()
 					yawMagToInputs[baseIntendedYaw][intendedMag] = stick;
 			}
 
-			inputsToYawMag[stickX][stickY] = {baseIntendedYaw, intendedMag};
+			inputsToYawMag[int8_t(stickX)][int8_t(stickY)] = {baseIntendedYaw, intendedMag};
 		}
 	}
 	return {yawMagToInputs, inputsToYawMag};
@@ -332,7 +332,7 @@ int M64::load()
 	try
 	{
 		f.seekg(0, std::ios::end);
-		int length = f.tellg();
+		int length = int(f.tellg());
 		if (length == 0)
 		{
 			std::cerr << "empty M64\n";
@@ -423,7 +423,7 @@ int M64::save(long initFrame)
 
 			// number of inputs
 			f.seekp(0x18, std::ios_base::beg);
-			uint32_t samples = lastFrame + 1;
+			uint32_t samples = uint32_t(lastFrame + 1);
 			f.write(reinterpret_cast<char*>(&samples), sizeof(uint32_t));
 
 			// m64 type
